+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Removed
 
+## [0.9.11] - 2025-09-20
 
 
 ## [0.9.10] - 2025-09-19
@@ -18,6 +19,7 @@
 - Created `validate_args_number` in `data_models.py` to check and validade excessive arguments (`*args` and `**kwargs`)
 - Updated all class methods in `Converter` to accept `*args` and `**kwargs` arguments, which will be used to catch `TypeError` in situations where user inputs more arguments than required when using program thought API
 - Created `tests` directory to store all test files
+- Created `test_utils.py` file to test functions on `utils.py` file
 
 
 ## [0.9.9] - 2025-09-18